Update: Put WriteDataSet into own class. Update: Replaced if-else with switch to make code easier to read and maintain. Update: Open/Create and Save Spreadsheet got own functions and are inserted into another place to separate IO Operations from memory operations.
